include number of cpu on the P2P exchanged data

+++ autodiscovery/pkg/gossip/service.go
@@ -1,11 +1,16 @@
 package gossip
 
 import (
+	"errors"
 	"log"
 	"time"
 
 	discovery "github.com/uniris/uniris-core/autodiscovery/pkg"
 	"github.com/uniris/uniris-core/autodiscovery/pkg/monitoring"
+)
+
+var (
+	ErrNotFoundOnUnreachableList = errors.New("cannot found the peer in the unreachableKeys list")
 )
 
 //Service is the interface that provide gossip methods
@@ -15,7 +20,9 @@ type Service interface {
 
 //Notifier is the interface that provides methods to notify gossip discovery
 type Notifier interface {
-	Notify(discovery.Peer) error
+	NotifyDiscoveries(discovery.Peer) error
+	NotifyReachable(pubk string) error
+	NotifyUnreachable(pubk string) error
 }
 
 type service struct {
@@ -134,22 +141,47 @@ func (s service) spread(init discovery.Peer, seeds []discovery.Seed, dChan chan<
 func (s service) handleCycleReachables(c *Cycle, eChan chan<- error) {
 	for p := range c.result.reaches {
 		log.Printf("Gossip reached peer: %s", p.Endpoint())
-		//Remove the target from the unreachable list if it is
-		if err := s.repo.RemoveUnreachablePeer(p.Identity().PublicKey()); err != nil {
+
+		err := s.repo.ContainsUnreachableKey(p.Identity().PublicKey())
+		if err != nil && err != ErrNotFoundOnUnreachableList {
 			eChan <- err
 			return
 		}
+		if err == nil {
+			//Remove the target from the unreachable list if it is
+			if err := s.repo.RemoveUnreachablePeer(p.Identity().PublicKey()); err != nil {
+				eChan <- err
+				return
+			}
+			//Notify for the reachable peer
+			if err := s.notif.NotifyReachable(p.Identity().PublicKey()); err != nil {
+				eChan <- err
+				return
+			}
+		}
+
 	}
 }
 
 func (s service) handleCycleUnreachables(c *Cycle, uChan chan<- discovery.Peer, eChan chan<- error) {
 	for p := range c.result.unreachables {
 		log.Printf("Gossip unreached peer: %s", p.Endpoint())
-		if err := s.repo.SetUnreachablePeer(p.Identity().PublicKey()); err != nil {
-			eChan <- err
-			return
+		err := s.repo.ContainsUnreachableKey(p.Identity().PublicKey())
+		if err != nil {
+			if err != ErrNotFoundOnUnreachableList {
+				eChan <- err
+				return
+			}
+			if err := s.repo.SetUnreachablePeer(p.Identity().PublicKey()); err != nil {
+				eChan <- err
+				return
+			}
+			if err := s.notif.NotifyUnreachable(p.Identity().PublicKey()); err != nil {
+				eChan <- err
+				return
+			}
+			uChan <- p
 		}
-		uChan <- p
 	}
 }
 
@@ -171,7 +203,7 @@ func (s service) handleCycleDiscoveries(c *Cycle, dChan chan<- discovery.Peer, e
 		}
 
 		//Notify the new discovery
-		if err := s.notif.Notify(p); err != nil {
+		if err := s.notif.NotifyDiscoveries(p); err != nil {
 			eChan <- err
 			return
 		}
